backend/requirements.txt
# Core Django
Django==5.2.9

# REST Framework
djangorestframework==3.16.1

# JWT Authentication
djangorestframework-simplejwt==5.5.1

# CORS Headers
django-cors-headers==4.9.0

# Environment Variables
python-dotenv==1.2.1

# PostgreSQL Database Driver
psycopg2-binary==2.9.10

# Production Server (Optional)
gunicorn==23.0.0